added minimal sms mod
very minimal so far; it just adds the phone_number field to the User
model and the the /me/edit page. but it took me a long time to figure
out how to do that. :)

i'm not sure if config/locales/*.yml is the correct place to add the
strings, but i don't see any examples of how to include locales within
mods, so it'll have to do, for now.